+++ testmod-common/src/main/java/io/github/jamalam360/testmod/TestModClient.java
@@ -3,6 +3,8 @@ package io.github.jamalam360.testmod;
 import io.github.jamalam360.jamlib.api.events.core.EventResult;
 import io.github.jamalam360.jamlib.api.network.Network;
 import io.github.jamalam360.jamlib.api.pack.PackReloadListenerRegistry;
+import io.github.jamalam360.jamlib.client.api.command.ClientCommandBuilders;
+import io.github.jamalam360.jamlib.client.api.command.ClientCommandRegistrationEvent;
 import io.github.jamalam360.jamlib.client.api.events.ClientContainerRenderEvents;
 import io.github.jamalam360.jamlib.client.api.events.ClientLevelTickEvents;
 import io.github.jamalam360.jamlib.client.api.events.ClientMouseScrollEvents;
@@ -11,6 +13,7 @@ import io.github.jamalam360.testmod.network.PotatoPacket;
 import io.github.jamalam360.testmod.pack.TestReloadListener;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Minecraft;
+import net.minecraft.network.chat.Component;
 import net.minecraft.server.packs.PackType;
 import org.lwjgl.glfw.GLFW;
 
@@ -32,6 +35,13 @@ public class TestModClient {
 			}
 		});
         PackReloadListenerRegistry.register(PackType.CLIENT_RESOURCES, TestMod.id("test_client_resources"), new TestReloadListener("client_resources"));
+
+        ClientCommandRegistrationEvent.EVENT.listen((dispatcher, context) -> {
+            dispatcher.register(ClientCommandBuilders.literal("helloc").executes(c -> {
+                c.getSource().client$sendSuccess(Component.literal("Hello, client!"));
+                return 1;
+            }));
+        });
 
 		if (Flags.DEBUG_MOUSE_SCROLL_EVENTS) {
 			ClientMouseScrollEvents.ALWAYS.listen((m, a) -> {