+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - Add a resource reload listener registry.
 - Add container render events.
 - Add an option for registered keybinds to also trigger while in screens.
+- Add command registration events.